fix: 자동차 이름 에러메시지에 숫자로 받은 경우 에러 추가
