+++ README.md
@@ -42,10 +42,10 @@
 
 ### 에러
 
-- [ ] 에러메시지
-  - [ ] 사용자가 잘못된 값을 입력할 경우
-    - [ ] ‘[ERROR]’로 시작하는 메시지와 함께 `Error`를 발생시킨다.
-    - [ ] 애플리케이션은 종료되어야 한다.
+- [O] 에러메시지
+  - [O] 사용자가 잘못된 값을 입력할 경우
+    - [O] ‘[ERROR]’로 시작하는 메시지와 함께 `Error`를 발생시킨다.
+    - [O] 애플리케이션은 종료되어야 한다.
 
   - [O] 자동차 이름 입력, 횟수 이름 공통 에러
     - [O] 사용자가 아무것도 입력하지 않은 경우
@@ -54,6 +54,8 @@
   - [O] 1) 자동차 이름 관련 에러
     - [O] 6자 이상 입력하였을 시 에러 메시지가 출력된다.
       - [O] “[ERROR] 입력 글자수를 초과하였습니다. 5자 이내로 자동차 이름을 입력해주세요.”
+    - [O] 문자로 입력하지 않았을 시 에러 메시지가 출력된다.
+      - [O] “[ERROR] [ERROR] 문자가 아닙니다. 문자로 입력해주세요..”
 
   - [O] 2) 시도 횟수 관련 에러
     - [O] 숫자가 아닌 문자로 입력하였을 경우
