+++ src/App.js
@@ -1,20 +1,19 @@
 import { getCarNamesInput, getRacingCountInput } from './views/inputView.js';
-import { isCarNamesInputNotEmpty } from './errors/carNamesErrors.js';
+import { isCarNamesInputNotEmpty, isCarNamesInputTypeOfNum, isCarNamesInputValidatedLength } from './errors/carNamesErrors.js';
 import splitAndTrimCarName from './models/carNamesTrimmer.js';
-import validateCarNamesLength from './controllers/carNamesInputController.js';
-import { isRacingCountInputEmpty, isRacingCountTypeNumber } from './models/validations/racingCountValidator.js';
+import { isRacingCountInputEmpty, isRacingCountTypeNumber } from './errors/racingCountErrors.js';
 import findProgressionLength from './models/racingProgressionLength.js';
-
 
 class App {
   async run() {
     const carNamesInput = await getCarNamesInput();
     isCarNamesInputNotEmpty(carNamesInput);
+
     const spliteAndtrimmedCarName = splitAndTrimCarName(carNamesInput);
+    isCarNamesInputTypeOfNum(spliteAndtrimmedCarName);
 
-    if (validateCarNamesLength(spliteAndtrimmedCarName)) {
+    if (isCarNamesInputValidatedLength(spliteAndtrimmedCarName)) {
       const racingCountInput = await getRacingCountInput();
-
       isRacingCountInputEmpty(racingCountInput);
       isRacingCountTypeNumber(racingCountInput);
       findProgressionLength(spliteAndtrimmedCarName, racingCountInput);
