--- src/Models/validations/carNamesValidator.js
@@ -1,5 +1,0 @@
-const carNamesLengthValidator = trimmedEachCarName => {
-  return trimmedEachCarName.every(element => element.length < 6);
-};
-
-export { carNamesLengthValidator };
