+++ src/constants/errorMessages.js
@@ -1,8 +1,9 @@
 // 공통
 export const ERROR_MESSAGES = {
-  EMPTY_INPUT: '[ERROR] 아무것도 입력하지 않으셨습니다. ', // 공통
-  OVERLENGTH_INPUT: '[ERROR] 입력 글자수를 초과하였습니다. 5자 이내로 자동차 이름을 입력해주세요. ', // 자동차 이름
-  INVALID_NUMBER: '[ERROR] 숫자가 아닙니다. 숫자로 입력해 주세요. ', // 자동차 이름
+  EMPTY_INPUT: '[ERROR] 아무것도 입력하지 않으셨습니다. ', 
+  OVERLENGTH_INPUT: '[ERROR] 입력 글자수를 초과하였습니다. 5자 이내로 자동차 이름을 입력해주세요. ', 
+  INVALID_STRING: '[ERROR] 문자가 아닙니다. 문자로 입력해주세요.',
+  INVALID_NUMBER: '[ERROR] 숫자가 아닙니다. 숫자로 입력해 주세요. ', 
 };
 
 // 코드 참고
