+++ src/errors/racingCountErrors.js
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGES } from '../../constants/errorMessages.js';
+import { ERROR_MESSAGES } from '../constants/errorMessages.js';
 
 const isRacingCountInputEmpty = racingCountInput => {
   if (racingCountInput === '') {
